block function

+++ src/modules/comments/comment.service.ts
@@ -4,12 +4,15 @@ import { CommentDto } from '../../modules/comments/dto/comment.dto';
 import { Inject, Injectable, HttpException } from '@nestjs/common';
 import { PROVIDERS, ERRORS } from '../../common/constants';
 import { Comments } from './comment.model';
+import { Posts } from '../posts/post.model';
 
 @Injectable()
 export class CommentService {
     constructor(
         @Inject(PROVIDERS.COMMENTS_PROVIDER)
         private readonly commentRepository: typeof Comments,
+        @Inject(PROVIDERS.POSTS_PROVIDER)
+        private readonly postRepository: typeof Posts,
     ) { }
 
     // create a comment for a post
@@ -18,7 +21,11 @@ export class CommentService {
         userId: number,
         commentDto: CommentDto
     ): Promise<any> {
-
+        //check if post exists
+        const post = await this.postRepository.findByPk(postId);
+        if (!post) {
+            throw new HttpException(ERRORS.POST.NOT_FOUND, 404)
+        }
         const comment = await this.commentRepository.create<Comments>({
             ...commentDto,
             postId,
@@ -62,12 +69,17 @@ export class CommentService {
         id: number
 
     ): Promise<any> {
-        const Ifcomment = await this.commentRepository.findOne({
-            where: { postId, userId, id }
-        })
+        const Ifcomment = await this.commentRepository.findByPk(id);
         if (!Ifcomment) {
-            throw new HttpException(ERRORS.COMMENT_NOT_FOUND,404)
+            throw new HttpException(ERRORS.COMMENT.NOT_FOUND, 404)
+
         }
+        const user = Ifcomment.userId;
+        
+        if (user !== userId) {
+            throw new HttpException(ERRORS.USER.NOT_AUTHORIZED, 404);
+        }   
+
         return await this.commentRepository.update(comment, { where: { postId, userId, updatedBy: userId, id } }
         )
     }
@@ -79,13 +91,16 @@ export class CommentService {
         id: number
 
     ): Promise<any> {
-        const Ifcomment = await this.commentRepository.findOne({
-            where: { postId, userId, id }
-        })
+     const Ifcomment = await this.commentRepository.findByPk(id);
         if (!Ifcomment) {
-            throw new HttpException(ERRORS.COMMENT_NOT_FOUND,404)
+            throw new HttpException(ERRORS.COMMENT.NOT_FOUND, 404)
 
         }
+        const user = Ifcomment.userId;
+        
+        if (user !== userId) {
+            throw new HttpException(ERRORS.USER.NOT_AUTHORIZED, 404);
+        }   
 
         return await this.commentRepository.destroy({
             where: { postId, userId, id }
